composer: verify file existence before trying to archive it

+++ FingerPrint/composer.py
@@ -64,7 +64,8 @@ class Archiver:
                 continue
             else:
                 temp_path = data_dir
-            if not os.path.exists(os.path.join(temp_path, os.path.basename(swf.path))):
+            if not os.path.exists(os.path.join(temp_path, os.path.basename(swf.path))) and \
+                os.path.exists(swf.path):
                 # do not copy twice the same file
                 shutil.copy2(swf.path, temp_path)
             for i in swf.links:
